requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
torch==2.6.0
pillow==11.1.0
pydantic==2.10.6
